Move PODynamicCheckoutEvent to delegate package
