+++ ui/src/main/kotlin/com/processout/sdk/ui/checkout/DynamicCheckoutInteractor.kt
@@ -11,8 +11,6 @@ import coil.request.ImageRequest
 import coil.request.ImageResult
 import com.processout.sdk.R
 import com.processout.sdk.api.dispatcher.POEventDispatcher
-import com.processout.sdk.api.model.event.PODynamicCheckoutEvent
-import com.processout.sdk.api.model.event.PODynamicCheckoutEvent.*
 import com.processout.sdk.api.model.event.PONativeAlternativePaymentMethodEvent
 import com.processout.sdk.api.model.event.PONativeAlternativePaymentMethodEvent.WillSubmitParameters
 import com.processout.sdk.api.model.request.*
@@ -55,10 +53,8 @@ import com.processout.sdk.ui.checkout.PODynamicCheckoutConfiguration.GooglePayCo
 import com.processout.sdk.ui.checkout.PODynamicCheckoutConfiguration.GooglePayConfiguration.CheckoutOption.DEFAULT
 import com.processout.sdk.ui.checkout.PODynamicCheckoutConfiguration.GooglePayConfiguration.TotalPriceStatus.ESTIMATED
 import com.processout.sdk.ui.checkout.PODynamicCheckoutConfiguration.GooglePayConfiguration.TotalPriceStatus.FINAL
-import com.processout.sdk.ui.checkout.delegate.DynamicCheckoutAlternativePaymentConfigurationRequest
-import com.processout.sdk.ui.checkout.delegate.DynamicCheckoutAlternativePaymentConfigurationResponse
-import com.processout.sdk.ui.checkout.delegate.DynamicCheckoutSavedPaymentMethodsConfigurationRequest
-import com.processout.sdk.ui.checkout.delegate.DynamicCheckoutSavedPaymentMethodsConfigurationResponse
+import com.processout.sdk.ui.checkout.delegate.*
+import com.processout.sdk.ui.checkout.delegate.PODynamicCheckoutEvent.*
 import com.processout.sdk.ui.napm.*
 import com.processout.sdk.ui.napm.PONativeAlternativePaymentConfiguration.*
 import com.processout.sdk.ui.savedpaymentmethods.POSavedPaymentMethodsConfiguration
